Initial auto-pull full implementation
