+++ src/functions/auto-pull/helpers/conf-prompter.js
@@ -53,15 +53,6 @@ async function getExcludedDirs(path) {
     type: 'checkbox',
     name: 'excludedDirs',
     message: 'Select repositories you would like not to enable auto-pulling for:',
-    choices: listDirs(path)
+    choices: require('../../../utils/fs/list-repo.js')(path)
   }])
-}
-
-function listDirs(path, parentDir = '') {
-  const files = fs.readdirSync(path, {withFileTypes: true})
-  const dirs = files.filter(file => file.isDirectory())
-  const repos = dirs.filter(dir => fs.existsSync(pathMod.join(path, dir.name, '.git'))).map(dir => pathMod.join(parentDir, dir.name))
-  const nonRepos = dirs.filter(dir => !fs.existsSync(pathMod.join(path, dir.name, '.git'))).map(dir => pathMod.join(parentDir, dir.name))
-  const subRepos = nonRepos.map(dir => listDirs(pathMod.join(path, pathMod.basename(dir)), dir)).flat(Infinity)
-  return [...repos, ...subRepos]
 }
