+++ src/functions/auto-pull/index.js
@@ -1,14 +1,17 @@
 module.exports = async (args) => {
   require('./helpers/check-logs.js')()
+  const Conf = require('conf')
+  const config = new Conf({
+    configName: 'auto-pull',
+    fileExtension: 'conf'
+  })
   if (['-c', '--config'].includes(args[0])) {
-    const Conf = require('conf')
-    const config = new Conf({
-      configName: 'auto-pull',
-      fileExtension: 'conf'
-    })
     config.store = await require('./helpers/conf-prompter.js')(args.splice(1))
   } else {
-    const info = await require('./helpers/info-prompter.js')()
-    console.log(info.name, info.age)
+    const path = config.get('path')
+    if (!path) {
+      throw new Error('No path defined in your auto-pull configuration. Please run git-assist auto-pull [-c, --config] in order to configurate your auto-pull utility.')
+    }
+    require('./helpers/auto-pull.js')(path, config.get('excludedDirs'))
   }
 }
